feat(profile): created settings-enum

+++ src/app/pages/main/profile/profile.page.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { PostitColorEnum } from 'src/app/models/enums/postit-color.enum';
+import { ProfileSettingsEnum } from 'src/app/models/enums/profile-settings.enum';
 import { PostitProxy } from 'src/app/models/proxies/postit.proxy';
 import { UserProxy } from 'src/app/models/proxies/user.proxy';
 import { NoteService } from 'src/app/modules/http-async/service/note.service';
@@ -17,13 +17,14 @@ export class ProfilePage {
   public postitArray: PostitProxy[] = [];
   public isLoading = false;
   public myUser: UserProxy;
+  public profileSettingsEnum: typeof ProfileSettingsEnum = ProfileSettingsEnum;
 
   constructor(
     public readonly note: NoteService,
     public readonly user: UserServise,
     public readonly helper: HelperService,
     public readonly router: Router,
-    ) { }
+  ) { }
 
   public async ionViewDidEnter(): Promise<void> {
     await this.loadMyNotes();
@@ -35,7 +36,7 @@ export class ProfilePage {
     const [notes, errorMessage] = await this.note.getMyPublicNotes();
     this.isLoading = false;
 
-    if (errorMessage) {return this.helper.showToast(errorMessage, 5_000);}
+    if (errorMessage) { return this.helper.showToast(errorMessage, 5_000); }
 
     this.postitArray = notes;
   }
@@ -45,25 +46,30 @@ export class ProfilePage {
     const [user, errorMessage] = await this.user.getUser(+localStorage.getItem(environment.keys.user));
     this.isLoading = false;
 
-    if (errorMessage) {return this.helper.showToast(errorMessage, 5_000);}
+    if (errorMessage) { return this.helper.showToast(errorMessage, 5_000); }
 
     this.myUser = user;
   }
 
-  public async clickConfigList(index: 0 | 1 | 2 | 3): Promise<void> {
-    switch (index) {
-      case 1 : {
-        break;
-      }
-      case 2 : {
-        break;
-      }
-      case 3 : {
-        localStorage.clear();
-        await this.router.navigate(['/login']);
-        break;
-      }
+  public async clickConfigList(selectedSettings: ProfileSettingsEnum): Promise<void> {
+    if (selectedSettings === ProfileSettingsEnum.EDIT_PROFILE) {
+      return this.helper.showToast('Em desenvolvimento');
+    }
+    if (selectedSettings === ProfileSettingsEnum.EDIT_PROFILE_PICTURE) {
+      return this.helper.showToast('Em desenvolvimento');
+    }
+    if (selectedSettings === ProfileSettingsEnum.ABOUT_US) {
+      return this.helper.showToast('Em desenvolvimento', 2_000);
+    }
+    if (selectedSettings === ProfileSettingsEnum.EXIT) {
+      localStorage.clear();
+      return void await this.router.navigate(['/login']);
     }
   }
 
+  public secondsToMilliseconds(seconds: number) {
+    const RELATION_SECONDS_MILLISECONDS = 1000;
+
+    return seconds * RELATION_SECONDS_MILLISECONDS;
+  }
 }
